* muzero cartpole work

+++ lzero/worker/muzero_collector.py
@@ -689,8 +689,10 @@ class MuZeroCollector(ISerialCollector):
                 stack_obs_tensor = prepare_observation(
                     stack_obs_array, self.policy_config.model.model_type
                 )
-                stack_obs_tensor = torch.from_numpy(stack_obs_tensor).to(
-                    self.policy_config.device
+                stack_obs_tensor = (
+                    torch.from_numpy(stack_obs_tensor)
+                    .to(self.policy_config.device)
+                    .float()
                 )
 
                 # ==============================================================
